requirements.txt
Flask==3.0.2
kafka-python-ng==2.2.2
openai==1.11.1
psycopg-binary==3.2.1
streamlit==1.38.0
python-dotenv
google-generativeai